requirements.txt
Django==2.2.9
djangorestframework==3.8.2
psycopg2==2.8.4
